requirements.txt
altwalker
Appium-Python-Client
